fixup! feat: wip multithreading
