+++ src/one_shot/mod.rs
@@ -1,4 +1,8 @@
+pub mod append_vec;
+pub mod appendable;
 pub mod bucket;
 pub mod matcher;
+pub mod matcher_parallel;
 
-pub use matcher::{match_list, match_list_parallel};
+pub use matcher::match_list;
+pub use matcher_parallel::match_list_parallel;
